refactor: update button to inline src styles using template
